requirements.txt
pybit==5.6.0
pandas==2.1.4
numpy==1.24.3
python-dotenv==1.0.0
websocket-client==1.6.4
asyncio==3.4.3
aiohttp==3.9.1
ccxt==4.1.77
